Deletes old tools; Moves contexts into libs; Moves storage funcs into storage/libs; moves images funcs into image/libs; changes imports for all funcs
