+++ openvair/modules/block_device/domain/fibre_channel/fibre_channel.py
@@ -16,7 +16,7 @@ Classes:
 """
 
 from openvair.libs.log import get_logger
-from openvair.modules.tools.utils import lip_scan as utils_lip_scan
+from openvair.modules.block_device.libs.utils import lip_scan as utils_lip_scan
 from openvair.modules.block_device.domain.base import BaseFibreChannel
 from openvair.modules.block_device.domain.exceptions import (
     LipScanError,
