+++ openvair/modules/virtual_machines/service_layer/services.py
@@ -38,8 +38,8 @@ from collections import namedtuple
 
 from openvair.libs.log import get_logger
 from openvair.libs.libvirt.domain import get_vms_state
-from openvair.modules.tools.utils import synchronized_session
 from openvair.modules.base_manager import BackgroundTasks, periodic_task
+from openvair.libs.context_managers import synchronized_session
 from openvair.modules.virtual_machines import config
 from openvair.libs.messaging.exceptions import (
     RpcCallException,
